Move get_bundles functionality to utils
