+++ tests/scale_test/utils/utils.py
@@ -2,6 +2,7 @@ import click
 import hvac
 import hvac.exceptions
 import json
+import logging
 import requests
 import uuid
 from requests_http_signature import HTTPSignatureAuth
@@ -362,6 +363,50 @@ def prepare_auth(lira_auth_dict):
             )
         )
     return valid_auth_dict
+
+
+def get_bundles(query_json, dss_url, output_format, replica):
+    """ Search for bundles in the HCA Data Storage Service using an elasticsearch query.
+
+    Args:
+        query_json (dict): Elasticsearch JSON query.
+        dss_url (str): URL for the HCA Data Storage Service.
+        output_format (str): Format of the query results, either "summary" for a list of UUIDs or "raw" to include'
+            the bundle JSON metadata.
+        replica (str): The cloud replica to search in, either "gcp" or "aws".
+
+    Returns:
+        list: List of dicts in the format { bundle_uuid: <uuid>, bundle_version: <version> }
+
+    """
+    search_url = '{}/v1/search?output_format={}&replica={}&per_page=500'.format(
+        dss_url.strip('/'), output_format, replica
+    )
+    headers = {'Content-type': 'application/json'}
+    response = requests.post(search_url, json=query_json, headers=headers)
+    results = response.json()['results']
+    total_hits = response.json()['total_hits']
+    logging.info('{} matching bundles found in {}'.format(total_hits, dss_url))
+    bundles = [format_bundle(r['bundle_fqid']) for r in results]
+
+    # The 'link' header refers to the next page of results to fetch. If there is no link header present,
+    # all results have been fetched.
+    # Example:
+    # link: <https://dss.dev.data.humancellatlas.org/v1/search?output_format=summary&replica=gcs&per_page=500&scroll_id=123>; rel="next"
+    link_header = response.headers.get('link', None)
+    while link_header:
+        next_link = link_header.split(';')[0]
+        next_url = next_link.strip('<').strip('>')
+        response = requests.post(next_url, json=query_json, headers=headers)
+        results = response.json()['results']
+        bundles.extend(format_bundle(r['bundle_fqid']) for r in results)
+        link_header = response.headers.get('link', None)
+    return bundles
+
+
+def format_bundle(bundle_fqid):
+    bundle_components = bundle_fqid.split('.', 1)
+    return {'bundle_uuid': bundle_components[0], 'bundle_version': bundle_components[1]}
 
 
 class AuthException(Exception):
